requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
djangorestframework==3.15.1
et-xmlfile==1.1.0
numpy==1.21.6
openpyxl==3.1.3
pandas==1.1.5
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
xlrd==2.0.1